auto (./check.c) 1 8

+++ check.c
@@ -207,7 +207,7 @@ static void maybe_commit(CC ctx, char* path, size_t words, size_t characters) {
 		uv_timer_stop(&ci.committer);
 		commit_now(path,words,characters);
 	} else {
-		printf("waiting %d\n",d);
+		printf("waiting %d\n",(int)d);
 		time_t now = time(NULL);
 		if(now + d > ci.next_commit) {
 			// keep pushing the timer ahead, so we change as much as possible before committing
